+++ src/decoder.js
@@ -566,6 +566,20 @@ module.exports = {
       deviceId: uint32(bytes, 4),
       ok: bool(bytes, 8)
     }
+  },
+
+  /**
+    * Decodes the response to a set-super-passwords request (function code 0x8x).
+    *
+    * @param {buffer}   buffer  64 byte NodeJS buffer
+    *
+    * @param {object}   Decoded set-super-passwords response object
+    */
+  SetSuperPasswords: function (bytes) {
+    return {
+      deviceId: uint32(bytes, 4),
+      ok: bool(bytes, 8)
+    }
   }
 }